wrap up auth stuff

+++ routes/web.php
@@ -17,7 +17,9 @@ use App\Http\Controllers\ListingController;
 Route::get('/',[ListingController::class, 'index']);
 
 //show create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
+
+Route::get('/listings/manage',[ListingController::class,'manage'])->middleware('auth');
 
 //show single listing
 Route::get('/listings/{listing}',[ListingController::class, 'show']);
@@ -26,13 +28,23 @@ Route::get('/listings/{listing}',[ListingController::class, 'show']);
 Route::post('/listings',[ListingController::class,'store']);
 
 //show edit form
-Route::get('/listings/{listing}/edit',[ListingController::class,'edit']);
+Route::get('/listings/{listing}/edit',[ListingController::class,'edit'])->middleware('auth');
 
 //submit to update
-Route::put('/listings/{listing}',[ListingController::class,'update']);
+Route::put('/listings/{listing}',[ListingController::class,'update'])->middleware('auth');
 
 //submit to delete
-Route::delete('/listings/{listing}',[ListingController::class,'destroy']);
+Route::delete('/listings/{listing}',[ListingController::class,'destroy'])->middleware('auth');
+
 
 //show register form
-Route::get('/register',[UserController::class,'create']);
+Route::get('/register',[UserController::class,'create'])->middleware('guest');
+
+Route::post('/users',[UserController::class,'store']);
+
+//show login form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+Route::post('/logout', [UserController::class,'logout'])->middleware('auth');
+
+Route::post('/users/login', [UserController::class, 'authenticate']);
